Move tests to their own project
