+++ Gibe.Navigation.Umbraco.Tests/Gibe.Navigation.Umbraco.Tests.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net10.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <IsPackable>false</IsPackable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Gibe.Navigation.Umbraco\Gibe.Navigation.Umbraco.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.NET.Test.Sdk" Version="18.0.1" />
+      <PackageReference Include="NUnit" Version="4.4.0" />
+      <PackageReference Include="NUnit3TestAdapter" Version="6.1.0" />
+      <PackageReference Include="NUnit.Analyzers" Version="4.7.0" />
+      <PackageReference Include="Moq" Version="4.20.72" />
+    </ItemGroup>
+
+</Project>
