+++ Gibe.Navigation.Umbraco/Gibe.Navigation.Umbraco.csproj
@@ -20,11 +20,6 @@
 
     <ItemGroup>
       <PackageReference Include="Umbraco.Cms.Core" Version="17.5.3" />
-      <PackageReference Include="Microsoft.NET.Test.Sdk" Version="18.0.1" PrivateAssets="all" />
-      <PackageReference Include="NUnit" Version="4.4.0" PrivateAssets="all" />
-      <PackageReference Include="NUnit3TestAdapter" Version="6.1.0" PrivateAssets="all" />
-      <PackageReference Include="NUnit.Analyzers" Version="4.7.0" PrivateAssets="all" />
-      <PackageReference Include="Moq" Version="4.20.72" PrivateAssets="all" />
     </ItemGroup>
 
 </Project>
